updated dtype in dino attention quantile

+++ colab_server_pkg/feature_extractor.py
@@ -39,10 +39,10 @@ def get_batch_dino_attn_maps(frames_list: list[np.ndarray]) -> torch.Tensor:
         attn_norm_batch = (attn_batch - attn_mins) / (attn_maxs - attn_mins + 1e-8)
 
         # Apply vectorized 70th percentile thresholding per batch item and scale by 0.8
-        attn_flat = attn_norm_batch.view(B, -1)
+        attn_flat = attn_norm_batch.view(B, -1).float()
         p70 = torch.quantile(attn_flat, 0.70, dim=-1, keepdim=True).unsqueeze(
             -1
-        )  # Shape [B, 1, 1]
+        ).to(attn_norm_batch.dtype)  # Shape [B, 1, 1]
         attn_norm_batch = (
             torch.where(
                 attn_norm_batch >= p70,
